pyproject.toml
[project]
name = "wallet-backend-test"
version = "0.1.0"
description = ""
authors = [
    {name = "Daniil Pautov",email = "danpautoff@yandex.ru"}
]
license = {text = "MIT"}
readme = "README.md"
requires-python = "3.12.*"
dependencies = [
    "django (>=5.2.1,<6.0.0)",
    "djangorestframework (>=3.16.0,<4.0.0)",
    "python-dotenv (>=1.1.0,<2.0.0)",
    "drf-spectacular (>=0.28.0,<0.29.0)"
]
package-mode = false


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.group.dev.dependencies]
flake8 = "^7.2.0"
isort = "^6.0.1"
black = "^25.1.0"
pre-commit = "^4.2.0"

